IntPtr should be long in C

+++ NativeAOT.CodeGenerator/Types/BuiltIn/IntPtrTypeDescriptor.cs
@@ -10,7 +10,7 @@ public partial class BuiltInTypeDescriptors
                 "nint.Zero",
                 new() {
                     { CodeLanguage.CSharpUnmanaged, "nint" },
-                    { CodeLanguage.C, "int" },
+                    { CodeLanguage.C, "long" },
                     { CodeLanguage.Swift, "Int" }
                 }
             );
